feat(cangjie_runtime): sync from upstream v1.2.0-alpha.17

+++ stdlib/libs/std/fs/native/file_system_unix.c
@@ -565,7 +565,7 @@ extern bool CJ_FS_CreateTempDir(char* path)
 }
 
 /* Copy symbolic link */
-extern int CJ_FS_CopyLink(char* linkName1, char* linkName2)
+extern int8_t CJ_FS_CopyLink(char* linkName1, char* linkName2)
 {
     char buf[MAX_PATH_LEN] = "";
     ssize_t len = readlink(linkName1, buf, sizeof(buf) - 1);
